refactor: modular decomposition — useTerminals, self-managed CommandPalette
- Extract terminal session state (list, activeId, themes, commands)
  from App.tsx into useTerminals.ts (singleton hook pattern)
- Move Cmd/Ctrl+K shortcut + open/close into CommandPalette.tsx
  (always mounted, owns its behavior)
- Move wsStatus signal into rpc.ts (owns the WebSocket connection)
- Rename registry.ts → terminals.ts (clearer name)
- Add SolidJS patterns section to CLAUDE.md
- Open #46 for future refactoring opportunities

App.tsx is now a thin layout shell (~80 lines, down from ~190).
